Implements Neon SIMD
This adds support for the NEON SIMD instructions. They are still nightly
only and some instructions are missing, but they are close to finishing
implementing all of them, so stabilization shouldn't be too far off.
There's a few semantic differences between the different targets which
leaves some open questions in the meantime.